Include utils into the engine
