+++ namui/src/type.ts
@@ -9,6 +9,7 @@ import {
   Canvas,
   Font,
   InputRect,
+  Color,
 } from "canvaskit-wasm";
 import { BuildErrorNotifier } from "./build/BuildErrorNotifier";
 import { BuildServerConnection } from "./build/BuildServerConnection";
@@ -273,6 +274,48 @@ export const Convert = {
   },
   xywhToCanvasKit(rect: XywhRect): InputRect {
     return CanvasKit.XYWHRect(rect.x, rect.y, rect.width, rect.height);
+  },
+  ColorToHsl(color: Color) {
+    const [r, g, b, a] = color;
+    const normalizedR = r || 0;
+    const normalizedG = g || 0;
+    const normalizedB = b || 0;
+    const max = Math.max(normalizedR, normalizedG, normalizedB);
+    const min = Math.min(normalizedR, normalizedG, normalizedB);
+    const delta = max - min;
+
+    let hue = 0;
+    if (delta !== 0) {
+      switch (max) {
+        case normalizedR: {
+          hue = (normalizedG - normalizedB) / delta;
+          break;
+        }
+        case normalizedG: {
+          hue = (normalizedB - normalizedR) / delta + 2;
+          break;
+        }
+        case normalizedB: {
+          hue = (normalizedR - normalizedG) / delta + 4;
+          break;
+        }
+        default: {
+          throw new Error("Can not calculate hue.");
+        }
+      }
+    }
+    hue = (hue < 0 ? hue + 6 : hue) / 6;
+
+    const lightness = (max + min) / 2;
+    const saturation =
+      delta === 0 ? 0 : delta / (1 - Math.abs(2 * lightness - 1));
+
+    return {
+      hue,
+      saturation,
+      lightness,
+      alpha: a || 0,
+    };
   },
 };
